+++ src/frontend/admin/src/services/factories/orders/index.ts
@@ -74,9 +74,9 @@ const build = (state?: OrderStatesEnum): Order => {
       children: [],
     },
     payment_schedule: [
-      orderPayment("6/27/2024", totalOrder / 3),
-      orderPayment("7/27/2024", totalOrder / 3),
-      orderPayment("8/27/2024", totalOrder / 3),
+      orderPayment("2024-06-27T00:00:00.000Z", totalOrder / 3),
+      orderPayment("2024-07-27T00:00:00.000Z", totalOrder / 3),
+      orderPayment("2024-08-27T00:00:00.000Z", totalOrder / 3),
     ],
     credit_card: CreditCardFactory(),
     has_waived_withdrawal_right: faker.datatype.boolean(),