+++ src/frontend/admin/src/utils/dates.tsx
@@ -1,10 +1,9 @@
 import { getDjangoLang } from "@/utils/lang";
-import { isTestEnv } from "@/utils/testing";
 
 export const formatShortDate = (isoDate: string): string => {
   return new Intl.DateTimeFormat(getDjangoLang(), {
     dateStyle: "short",
     timeStyle: "short",
-    ...(isTestEnv() ? { timeZone: "UTC" } : {}),
+    timeZone: "UTC",
   }).format(new Date(isoDate));
 };